Fix an issue where the dates returned could contain NaN. Instead  these dates are now optionals.
The original Javascript library used acos() without checking the range of the input. All implementations based on the original Javascript library have the same issue. See:

https://github.com/mourner/suncalc/issues/134
https://github.com/mourner/suncalc/issues/65
https://github.com/shanus/flutter_suncalc/issues/1

This issue can be reproduced if the latitude/longitude is set for example to 89.0/179.0 or 84.0/111.0

The solution implemented here is based on the fix suggested in the Flutter library: https://github.com/shanus/flutter_suncalc/issues/1